Eliminate redundant utility class. Simplify packing strings.
